+++ packages/ai/src/prompts/index.ts
@@ -1,7 +1,9 @@
 import { IAudioPrompt } from './audio';
-import { IChatPrompt } from './chat';
+import { IChatPrompt } from './chat-types';
 
 export type Prompt = IChatPrompt | IAudioPrompt;
 
-export * from './chat';
 export * from './audio';
+export * from './chat';
+export * from './chat-types';
+
