+++ packages/openai/src/chat.ts
@@ -48,7 +48,7 @@ export type AzureOpenAIChatModelOptions = OpenAIChatModelOptions & {
   azureADTokenProvider?: () => Promise<string>;
 };
 
-export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
+export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams, OpenAI.Chat.ChatCompletionMessage> {
   private readonly _openai: OpenAI;
   private readonly _log: ILogger;
 
@@ -58,33 +58,33 @@ export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
     this._openai =
       'endpoint' in options
         ? new AzureOpenAI({
-            apiKey: options.apiKey,
-            apiVersion: options.apiVersion,
-            endpoint: options.endpoint?.replace(/\/$/, ''),
-            deployment: options.model,
-            azureADTokenProvider: options.azureADTokenProvider,
-            baseURL: options.baseUrl?.replace(/\/$/, ''),
-            organization: options.organization,
-            project: options.project,
-            defaultHeaders: options.headers,
-            fetch: options.fetch,
-            timeout: options.timeout,
-          })
+          apiKey: options.apiKey,
+          apiVersion: options.apiVersion,
+          endpoint: options.endpoint?.replace(/\/$/, ''),
+          deployment: options.model,
+          azureADTokenProvider: options.azureADTokenProvider,
+          baseURL: options.baseUrl?.replace(/\/$/, ''),
+          organization: options.organization,
+          project: options.project,
+          defaultHeaders: options.headers,
+          fetch: options.fetch,
+          timeout: options.timeout,
+        })
         : new OpenAI({
-            apiKey: options.apiKey,
-            baseURL: options.baseUrl?.replace(/\/$/, ''),
-            organization: options.organization,
-            project: options.project,
-            defaultHeaders: options.headers,
-            fetch: options.fetch,
-            timeout: options.timeout,
-          });
+          apiKey: options.apiKey,
+          baseURL: options.baseUrl?.replace(/\/$/, ''),
+          organization: options.organization,
+          project: options.project,
+          defaultHeaders: options.headers,
+          fetch: options.fetch,
+          timeout: options.timeout,
+        });
   }
 
   async send(
     input: Message,
     options: ChatSendOptions<ChatCompletionCreateParams> = {}
-  ): Promise<ModelMessage> {
+  ): Promise<ModelMessage<OpenAI.Chat.ChatCompletionMessage>> {
     const memory = options.messages || new LocalMemory();
     await memory.push(input);
 
@@ -137,13 +137,13 @@ export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
           Object.keys(options.functions || {}).length === 0
             ? undefined
             : Object.values(options.functions || {}).map((fn) => ({
-                type: 'function',
-                function: {
-                  name: fn.name,
-                  description: fn.description,
-                  parameters: fn.parameters,
-                },
-              })),
+              type: 'function',
+              function: {
+                name: fn.name,
+                description: fn.description,
+                parameters: fn.parameters,
+              },
+            })),
         messages: messages.map((message) => {
           if (message.role === 'model') {
             return {
@@ -179,15 +179,15 @@ export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
                 typeof message.content === 'string'
                   ? message.content
                   : message.content.map((p) => {
-                      if (p.type === 'image_url') {
-                        return {
-                          type: p.type,
-                          image_url: { url: p.image_url },
-                        };
-                      }
+                    if (p.type === 'image_url') {
+                      return {
+                        type: p.type,
+                        image_url: { url: p.image_url },
+                      };
+                    }
 
-                      return p;
-                    }),
+                    return p;
+                  }),
             };
           }
 
@@ -265,7 +265,7 @@ export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
         }
       }
 
-      const modelMessage: ModelMessage = {
+      const modelMessage: ModelMessage<OpenAI.Chat.ChatCompletionMessage> = {
         role: 'model',
         audio: message.audio || undefined,
         content: message.content || undefined,
@@ -275,6 +275,7 @@ export class OpenAIChatModel implements IChatModel<ChatCompletionCreateParams> {
           name: call.function.name,
           arguments: JSON.parse(call.function.arguments || '{}'),
         })),
+        raw: message,
       };
 
       if (message.tool_calls && message.tool_calls.length > 0) {
